python/trezorctl: only accept json for sign-tx, document the format
